feat(*): DetailChatViewController-UserInputSection 구현
유저 input section 부분 구현 및 autolayout적용
